fixed bugs about PipeReader (#1476)
* skip PipeReader in IO

* skip PipeReader in IO
test=develop

* fixed bugs about PipeReader

+++ doc/fluid/api/gen_doc.py
@@ -83,7 +83,8 @@ class DocGenerator(object):
         else:
             self.stream.close()
             path = os.getcwd()+"/fluid/"+name+".rst"
-            os.remove(path)
+            if name != "PipeReader":
+                os.remove(path)
 
     def print_class(self, name):
         self._print_ref_(name)
